chore(cleanup): temuan review pasca-v0.20.0 — relokasi FilterCombobox, barrel dirapikan, test scope, TLS strict — #217
Review 4 commit sejak rilis: FilterCombobox pindah ke shared/filter-combobox.tsx
(primitif generik, bukan bagian district-group-filter); barrel shared/index.ts
kembali ke pengecualian terdokumentasi (4 dead type exports dihapus, semua
konsumen impor per-path); DistrictFilterOption/GroupFilterOption dipakai 4 list
client menggantikan interface lokal duplikat; +4 test scope getDistrictsForSelect
(efek penyempitan form Lembaga dipertahankan by design); clamp angka negatif
tooltip matriks cakupan; rejectUnauthorized:false warisan SIGAP dihapus (kedua
host baru chain-valid); label lembaga terpilih robust utk init luar-distrik.
