+++ src/app/(admin)/admin/master-data/production/components/production-list-client.tsx
@@ -11,9 +11,13 @@ import { toast } from "sonner";
 import {
   TableActions,
   DataTable,
-  DistrictGroupFilter,
   type DataTableColumn,
 } from "@/components/shared";
+import {
+  DistrictGroupFilter,
+  type DistrictFilterOption,
+  type GroupFilterOption,
+} from "@/components/shared/district-group-filter";
 import {
   Select,
   SelectContent,
@@ -45,8 +49,8 @@ interface ProductionRecord {
 
 interface Props {
   initialRecords: ProductionRecord[];
-  farmerGroups: { id: string; name: string; code?: string | null; districtId: string }[];
-  districts: { id: string; name: string }[];
+  farmerGroups: GroupFilterOption[];
+  districts: DistrictFilterOption[];
   permissions: string[];
   isSuperAdmin: boolean;
 }
