+++ src/app/(admin)/admin/report/production/production-report-client.tsx
@@ -7,7 +7,7 @@ import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Badge } from "@/components/ui/badge";
 import { Input } from "@/components/ui/input";
-import { FilterCombobox } from "@/components/shared/district-group-filter";
+import { FilterCombobox } from "@/components/shared/filter-combobox";
 import { getFarmerGroupsForProductionReport, getProductionReport } from "@/server/actions/report";
 import type { ProductionReportResult } from "@/types/report";
 import { formatPeriodLabel, enumeratePeriods, PRODUCTION_REPORT_MAX_MONTHS } from "@/lib/report-production";
